// sourcecode/BTrak.Packages/snova-admin-module/admin-module-components/admin-components/src/lib/admin-module/models/hr-models/webhook-model.ts
import { SearchCriteriaInputModelBase } from '../searchCriteriaInputModelBase';

export class WebHookModel extends SearchCriteriaInputModelBase {
    webHookId: string;
    webHookName: string;
    webHookUrl: string;
    createdDateTime: Date;
    createdOn: string;
    inActiveDateTime: Date;
    timeStamp: any;
    totalCount: number;
}